Moved ProgressDemo code to separate controller; Fixes #12;
