+++ src/DeathStarImperator.UI/Controllers/HomeController.cs
@@ -1,6 +1,4 @@
-﻿using System.Threading;
-using System.Web.Mvc;
-using DeathStarImperator.UI.Models;
+﻿using System.Web.Mvc;
 
 namespace DeathStarImperator.UI.Controllers
 {
@@ -29,32 +27,6 @@ namespace DeathStarImperator.UI.Controllers
         public ActionResult Chat()
         {
             return View();
-        }
-
-        public ActionResult ProgressDemo()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        public ActionResult DoJob()
-        {
-            var job = JobManager.Instance.DoJobAsync(j =>
-            {
-                for (var progress = 0; progress <= 100; progress++)
-                {
-                    Thread.Sleep(200);
-                    j.ReportProgress(progress);
-                }
-            });
-
-            return Json(new
-            {
-                JobId = job.Id,
-                Progress = job.Progress
-            });
-        }
+        }        
     }
-
-    
 }
