+++ src/DeathStarImperator.UI/Views/Shared/_Layout.cshtml
@@ -40,7 +40,7 @@
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav navbar-right">
-                    <li>@Html.ActionLink("PROGRESS", "ProgressDemo", "Home")</li>
+                    <li>@Html.ActionLink("PROGRESS", "ProgressDemo", "ProgressDemo")</li>
                     <li>@Html.ActionLink("CHAT", "Chat", "Home")</li>
                 </ul>
             </div>
